+++ backend/src/main/java/com/sustainablefarm/dto/request/RawIntakeUpdateRequest.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -27,7 +28,7 @@ public class RawIntakeUpdateRequest {
     private LocalDate intakeDate;
 
     @Positive(message = "Received quantity must be positive")
-    private Double receivedQuantityKg;
+    private BigDecimal receivedQuantityKg;
 
     private MangoVariety receivedVariety;
